fix(types): resolve all mypy errors across 5 files
- models.py: handle str | None for azure_endpoint/deployment, fix
  Anthropic content block union-attr, handle refusal_keywords None
- likert/evaluator.py, ranking/evaluator.py: use BaseException instead
  of Exception in gather result filtering so mypy narrows correctly
- ranking/scorer.py: cast numpy floating to float for _sigmoid
- report/charts.py: add targeted type: ignore for matplotlib Collection
  and pandas unstack fill_value stub limitations

+++ src/political_bias/ranking/scorer.py
@@ -68,7 +68,7 @@ def compute_vote_shares(
         for c, avg_r in avg_ranks.items():
             z = (avg_r - mu) / sigma
             # Lower rank is better, so negate z
-            raw_scores[c] = _sigmoid(-z * k)
+            raw_scores[c] = _sigmoid(float(-z * k))
 
         # Apply position bias corrections if available (per-run, by label)
         if position_bias_corrections and model_id in position_bias_corrections:
